Patch product view implementation

+++ products/views.py
@@ -33,7 +33,7 @@ def API_products_create_list_view(request):
         data = serialize(
             "json",
             products,
-            fields=("name", "description", "author_user", "time_created"),
+            fields=("name", "description", "author_user", "time_last_updated"),
         )
         return HttpResponse(data, content_type="application/json", status=200)
 
@@ -50,7 +50,7 @@ def API_products_create_list_view(request):
         data = serialize(
             "json",
             [new_product],
-            fields=("pk", "name", "description", "user", "time_created"),
+            fields=("pk", "name", "description", "user", "time_last_updated"),
         )
         return HttpResponse(data, content_type="application/json", status=201)
 
@@ -63,11 +63,33 @@ def API_products_create_list_view(request):
 @login_required
 def API_products_update_delete_view(request, product):
     """Get API entry point for product partial update and product delete."""
-    if request.method == "DELETE":
+    if request.method == "PATCH":
         product_content = Product.objects.get(pk=product)
-        product_content.delete()
 
-        return HttpResponse("Object deleted.", status=204)
+        if product_content:
+            body = json.loads(request.body)
+
+            if body["name"]:
+                product_content.name = body["name"]
+            product_content.description = body["name"]
+            product_content.author_user = request.user
+
+            product_content.save()
+
+            return HttpResponse("Object updated.", status=204)
+
+        return HttpResponse("Object not found.", status=404)
+
+    elif request.method == "DELETE":
+        product_content = Product.objects.get(pk=product)
+
+        if product_content:
+            product_content.delete()
+
+            return HttpResponse("Object deleted.", status=204)
+
+        return HttpResponse("Object not found.", status=404)
+
     else:
         return HttpResponse(
             "Error 405 : This method is not allowed for this url.", status=405
